Moved exceptions to corresponding modules.
