+++ chatterbot/adapters/logic/closest_match.py
@@ -1,4 +1,3 @@
-from chatterbot.adapters.exceptions import EmptyDatasetException
 from .base_match import BaseMatchAdapter
 from fuzzywuzzy import process
 
@@ -25,7 +24,7 @@ class ClosestMatchAdapter(BaseMatchAdapter):
                 # Use a randomly picked statement
                 return 0, self.context.storage.get_random()
             else:
-                raise EmptyDatasetException
+                raise self.EmptyDatasetException()
 
         # Get the text of each statement
         text_of_all_statements = []
