+++ chatterbot/adapters/logic/logic_adapter.py
@@ -22,3 +22,11 @@ class LogicAdapter(Adapter):
         a confidence value and a statement as output.
         """
         raise self.AdapterMethodNotImplementedError()
+
+    class EmptyDatasetException(Exception):
+
+        def __init__(self, message="An empty collection of elements was received when at least one entry was expected."):
+            self.message = message
+
+        def __str__(self):
+            return self.message
