+++ chatterbot/adapters/storage/__init__.py
@@ -1,4 +1,4 @@
-from .storage import StorageAdapter
+from .storage_adapter import StorageAdapter
 from .jsondatabase import JsonDatabaseAdapter
 from .mongodb import MongoDatabaseAdapter
 from .twitter_storage import TwitterAdapter
